+++ src/Service/Comprehend/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - AWS enhancement: Documentation updates.
+- AWS api-change: Add reason and detail in `InvalidRequestException` exception
 
 ## 0.1.0